Added ARenderable scaling
Also messed with texture management. Since we set the destination
rectangle using the size of the sprite instead of the size of the
texture, some sprites were broken in the tester because they were set
from images and setting a texture from an image does not update the
source sprite size. Now setting an animation or a image texture on a
ARenderable will update the size.

Cleaned some places where the * of a pointer was not touching the type
name.

+++ engine/src/ATexture.cpp
@@ -41,13 +41,13 @@ ArrayList<ATexture*>* ATexture::SplitTexture(int splits)
 {
     ArrayList<ATexture*>* result = new ArrayList<ATexture*>(splits);
 
-    int splitWidth = this->textureSize.Width / splits;
+    int splitWidth = (int)this->textureSize.Width / splits;
 
     int splitStartX = 0;
     int splitStartY = 0;
     for(int i = 0; i < splits; i++)
     {
-        ATexture* split = this->GetSubTexture(splitStartX, splitStartY, splitWidth, this->textureSize.Height);
+        ATexture* split = this->GetSubTexture(splitStartX, splitStartY, splitWidth, (int)this->textureSize.Height);
 
         splitStartX += splitWidth;
